Made methods static, removed unnecessary interface.

+++ credit_card_prefix.cpp
@@ -6,18 +6,18 @@
 #include "fake_credit_card_generator.h"
 #include <time.h>
 
-long CreditCardPrefix::getDiscover() {
+short unsigned int CreditCardPrefix::getDiscover() {
 	srand (time(NULL));
 	int prefixes[] = {6011, 65, rand() % 622126 + 799, rand() % 644 + 5};
 	return prefixes[rand() % sizeof(prefixes)/sizeof(*prefixes)];
 }
-long CreditCardPrefix::getMasterCard() {
+short unsigned int CreditCardPrefix::getMasterCard() {
 	return rand() % 51 + 4;
 }
-long CreditCardPrefix::getVisa() {
+short unsigned int CreditCardPrefix::getVisa() {
 	return 4;
 }
-long CreditCardPrefix::getAmericanExpress() {
+short unsigned int CreditCardPrefix::getAmericanExpress() {
 	int prefixes[] = {34, 37};
 	return prefixes[rand() % sizeof(prefixes)/sizeof(*prefixes)];
 }
